Teach :Rake about spec for app/ files
If :Rake sees that RSpec is present and there is a corresponding spec/
file for the currently edited app/ file, it will now run that spec
instead of blindly running the Test::Unit file.

The implementation is less than ideal.  It blindly substitutes /app/ for
/spec/, which could easily have false positives, and fails to recognize
view specs.  It also ignores any Test::Unit tests (the preferred
behavior would be to yield to Test::Unit in a tie, or to run both).