Rproject checks directory existence before adding to project
When :Rproject was run, if any of the directories it added to
the project didn't exist, an error dialog would appear, if
guioptions did not include c. If guioptions did include c, it
would look like the process had hung, as it was waiting for
the user to clear the message at the bottom of the window,
which was invisible.

This patch checks for the existence of spec directories before
adding them to the project. It also checks for the existence
of the test directory, since RSpecers typically delete it.