Oracle dbext support fixes
The Oracle Rails adaptor is called 'oracle' rather than 'oci' (this
changed around the time of Rails 2, I think).  Also the Oracle adaptor
expects all the connection information to be in the 'database' config
value, not separately in host, port, database, etc. This is because
Oracle understands various ways of expressing this info, such as the
EZ connect string (host:port/database), a TNS name,
an LDAP name, etc.  The Rails adaptor justs passes the value of
'database' through to the Oracle client, which sorts all that out.
Similarly, dbext's Oracle support expects this stuff in one place,
the 'srvname' config option.  It doesn't seem to use the value of
'dbname' at all, so the fact that rails.vim still sets it doesn't
cause any harm.