Support custom "classifications" in config/editor.json
A classification is a category of file in your app: model, controller,
helper, and so forth.  You can now provide custom classification
definitions in your application's config/editor.json and they will be
automatically defined as navigation commands.  Example config:

    {
      "classifications": {
        "concern": {
          "prefix": "app/concerns/",
          "suffix": ".rb"
        },
        "fabricator": {
          "alternate": "app/models/%s.rb",
          "affinity": "model",
          "prefix": ["spec/fabricators/", "test/fabricators/"],
          "suffix": "_fabricator.rb"
        }
      }
    }

The "affinity" attribute lets the classification receive and provide
default arguments.  With this example, one could use :Rmodel in a
fabricator to jump to the model, and vice versa.

The "alternate" attribute provides the destination for the :A command.

Suggestions for alternative names for "classification" (and "affinity")
welcome.  "Type" is a confusing choice because Vim already has file
types.