Map Rails 4 testing directories to rake tasks
This adds the Rails 4 idiom of creating test directories named "models"
and "controllers" (in lieu of "unit" and "functional") to the pattern
matches when running `:Rake`:, selecting the appropriate rake task to
run.

This allows those doing Rails 4 development to run `:Rake` and avoid
falling down to the deprecated `test:recent`.