Add tutor with sample app
The best way to learn something is by doing.
Getting started with something new can often be intimidating,
and going through the documentation of a project does not leave you
with the muscle memory to be able to repeat what you have read.

This commit provides a vimtutor-like instructions and a sample Rails
application to help you practice your newly learned skills.