Be smarter about overriding 'includeexpr'
Between ever evolving naming conventions and the myriad of preprocessors
and template languages, it's hard to pin down a good rule of when
exactly this option should be set.  Let's try a heuristic based on the
old value.

Also, tie the gf family of maps to the setting of this option.